refactor(model): adaptação dos services para a nova estrutura com diretório model.
