+++ src/main/java/com/ufma/portifolium/service/ProfessorService.java
@@ -3,7 +3,7 @@ package com.ufma.portifolium.service;
 import java.util.List;
 import java.util.Optional;
 
-import com.ufma.portifolium.entities.Professor;
+import com.ufma.portifolium.model.entities.Professor;
 import com.ufma.portifolium.repository.ProfessorRepository;
 import com.ufma.portifolium.service.exceptions.CadastroException;
 import com.ufma.portifolium.utils.Utils;
