+++ src/main/java/com/ufma/portifolium/service/UsuarioService.java
@@ -3,7 +3,7 @@ package com.ufma.portifolium.service;
 import java.util.List;
 import java.util.Optional;
 
-import com.ufma.portifolium.entities.Usuario;
+import com.ufma.portifolium.model.entities.Usuario;
 import com.ufma.portifolium.repository.UsuarioRepository;
 import com.ufma.portifolium.service.exceptions.AutenticacaoException;
 import com.ufma.portifolium.service.exceptions.CadastroException;
