+++ src/test/java/com/ufma/portifolium/PortifoliumApplicationTests.java
@@ -1,5 +1,6 @@
 package com.ufma.portifolium;
 
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
@@ -8,6 +9,7 @@ class PortifoliumApplicationTests {
 
 	@Test
 	void contextLoads() {
+		Assertions.assertTrue(true);
 	}
 
 }
